pyproject.toml
[project]
name = "protovoice"
version = "0.2.0"
description = "Duplex voice agent on Pipecat — Whisper STT + Qwen LLM + Fish/Kokoro TTS"
requires-python = ">=3.11"
dependencies = [
    # Pipecat + transports
    "pipecat-ai[webrtc,silero,openai]>=1.0",
    # Local model runtimes
    "torch>=2.5",
    "transformers>=4.46",
    "accelerate",
    "kokoro>=0.9",
    "vllm>=0.18",
    # Web + I/O
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "httpx",
    "soundfile",
    "soxr",
    "numpy",
    "pyyaml",
    "ddgs",
]

[project.scripts]
protovoice = "app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
